Refactor account balance to Double and improve UI components
- Change account balance types from Int to Double across models and services
- Make TransactionCategory iconName optional with safe unwrapping
- Refactor AccountDetailViewModel with proper dependency injection initialization
- Update AccountDetailView to pass accountId during initialization
- Improve chat input view padding and placeholder positioning
- Hide tab bar in ChatRoomView for better UX
- Update TransactionParams with proper initialization and CodingKeys
- Remove commented category breakdown code from AccountSummaryCard
- Comment out Transactions tab temporarily in AppView

+++ FinancialApp/Model/Accounts/AccountModel.swift
@@ -7,9 +7,9 @@ struct AccountModel: Identifiable, Codable, Sendable, Equatable {
   let category: AccountCategory
   let currency: String
   let createdAt: Date?
-  let finalBalance: Int
+  let finalBalance: Double
   let accountNumber: String?
-  let initialBalance: Int
+  let initialBalance: Double
   var updatedAt: Date? = nil
   
   enum CodingKeys: String, CodingKey {
@@ -26,11 +26,11 @@ struct AccountModel: Identifiable, Codable, Sendable, Equatable {
   }
   
   // Computed properties
-  var balance: Int {
+  var balance: Double {
     finalBalance
   }
   
-  var balanceChange: Int {
+  var balanceChange: Double {
     finalBalance - initialBalance
   }
